// src/app/shared/core/interfaces/notemeta.ts
export interface Notemeta {
  __v: Number;
  _archiveplace: Number;
  _id: String;
  arrangeur: String;
  auffuerungs_jahr: String;
  aufnahme_url: String;
  bemerkungen: String;
  createdAt: String;
  demo_url: String;
  digital_analog: String;
  duration: String;
  flex: String;
  grad: String;
  group: String;
  jmr_aufnahme_url: String;
  jungmusik_fest: String;
  komponist: String;
  stil: String;
  title: String;
  updatedAt: String;
  verlag: String;
}